// src/app/pages/apps/unidades/interfaces/unidades.model.ts
export class Unidades {
  idUnidad?: number;
  tipoUnidad: string;
  eco: string;
  placas: string;
  numeroSerie: string;
  caratula: string;
  vigenciaInicio: string;
  vigenciaFin: string;
  inciso: string;
  cobertura: string;
  marca: string;
  modelo: string;
  anio: string;
  paisOrigen: string;
  marcaMotor: string;
  numeroMotor: string;
  ejes: string;
  llantas: string;
  largoFts: string;
  anchoFts: string;
  altoFts: string;
  pesoVehicular: string;
  pesoMaximo: string;
  tarjetaCirculacion: string;
  fisicoMecanica: string;
  fechaFM: string;
  categoria: string;
  emisionContaminantes: string;
  fechaEC: string;
  periodo: string;
  tagChihuahua: string;
  tagPase: string;
  autorizacionExpresa: string;
  permisoCirculacion: string;
  alias: string;
  capacidad: string;
  satelite: string;

  constructor(unidad: any) {
    this.idUnidad = unidad.idUnidad;
    this.tipoUnidad = unidad.tipoUnidad;
    this.eco = unidad.eco;
    this.placas = unidad.placas;
    this.numeroSerie = unidad.numeroSerie;
    this.caratula = unidad.caratula;
    this.vigenciaInicio = unidad.vigenciaInicio;
    this.vigenciaFin = unidad.vigenciaFin;
    this.inciso = unidad.inciso;
    this.cobertura = unidad.cobertura;
    this.marca = unidad.marca;
    this.modelo = unidad.modelo;
    this.anio = unidad.anio;
    this.paisOrigen = unidad.paisOrigen;
    this.marcaMotor = unidad.marcaMotor;
    this.numeroMotor = unidad.numeroMotor;
    this.ejes = unidad.ejes;
    this.llantas = unidad.llantas;
    this.largoFts = unidad.largoFts;
    this.anchoFts = unidad.anchoFts;
    this.altoFts = unidad.altoFts;
    this.pesoVehicular = unidad.pesoVehicular;
    this.pesoMaximo = unidad.pesoMaximo;
    this.tarjetaCirculacion = unidad.tarjetaCirculacion;
    this.fisicoMecanica = unidad.fisicoMecanica;
    this.fechaFM = unidad.fechaFM;
    this.categoria = unidad.categoria;
    this.emisionContaminantes = unidad.emisionContaminantes;
    this.fechaEC = unidad.fechaEC;
    this.periodo = unidad.periodo;
    this.tagChihuahua = unidad.tagChihuahua;
    this.tagPase = unidad.tagPase;
    this.autorizacionExpresa = unidad.autorizacionExpresa;
    this.permisoCirculacion = unidad.permisoCirculacion;
    this.alias = unidad.alias;
    this.capacidad = unidad.capacidad;
    this.satelite = unidad.satelite;
  }
}